AARects: XYRect, XZRect, YZRect to separate module
